docs: add http-client ADR
Create a first ADR to define the http-client, a core component that will
drive the core networking level of ladon.

This component will be resposible to interact via HTTP with the sites to
be analyzed, and will be responsible for the interaction with sites. Two
aspects are mainly taken into consideration:

- Site crawling policy, respecting robots.txt and rate limiter
- Monitoring and observability. This aspect is particularly important,
  because it's really important that every event is monitored and
traced for future analysis and statistics production.